Fix some nits
