+++ lock-api/src/main/java/com/palantir/lock/watch/LockWatch.java
@@ -26,7 +26,7 @@ import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 @JsonSerialize(as = ImmutableLockWatch.class)
 @JsonDeserialize(as = ImmutableLockWatch.class)
 public interface LockWatch {
-    LockWatch INVALID = ImmutableLockWatch.of(0L, false);
+    LockWatch INVALID = ImmutableLockWatch.of(-1L, false);
 
     @Value.Parameter
     long timestamp();
